m101: Update core/ and effects/
