+++ skia-safe/src/core/alpha_type.rs
@@ -1,0 +1,2 @@
+pub use skia_bindings::SkAlphaType as AlphaType;
+variant_name!(AlphaType::Premul, alpha_type_naming);
