+++ skia-safe/src/core/image_info.rs
@@ -1,72 +1,6 @@
-use crate::{prelude::*, ColorSpace, IPoint, IRect, ISize};
-use skia_bindings::{self as sb, SkColorInfo, SkColorType, SkImageInfo};
+use crate::{prelude::*, AlphaType, ColorSpace, ColorType, IPoint, IRect, ISize};
+use skia_bindings::{self as sb, SkColorInfo, SkImageInfo};
 use std::{fmt, mem};
-
-pub use skia_bindings::SkAlphaType as AlphaType;
-variant_name!(AlphaType::Premul, alpha_type_naming);
-
-#[derive(Copy, Clone, PartialEq, Eq, Hash, Debug)]
-#[repr(i32)]
-pub enum ColorType {
-    Unknown = SkColorType::kUnknown_SkColorType as _,
-    Alpha8 = SkColorType::kAlpha_8_SkColorType as _,
-    RGB565 = SkColorType::kRGB_565_SkColorType as _,
-    ARGB4444 = SkColorType::kARGB_4444_SkColorType as _,
-    RGBA8888 = SkColorType::kRGBA_8888_SkColorType as _,
-    RGB888x = SkColorType::kRGB_888x_SkColorType as _,
-    BGRA8888 = SkColorType::kBGRA_8888_SkColorType as _,
-    RGBA1010102 = SkColorType::kRGBA_1010102_SkColorType as _,
-    BGRA1010102 = SkColorType::kBGRA_1010102_SkColorType as _,
-    RGB101010x = SkColorType::kRGB_101010x_SkColorType as _,
-    BGR101010x = SkColorType::kBGR_101010x_SkColorType as _,
-    Gray8 = SkColorType::kGray_8_SkColorType as _,
-    RGBAF16Norm = SkColorType::kRGBA_F16Norm_SkColorType as _,
-    RGBAF16 = SkColorType::kRGBA_F16_SkColorType as _,
-    RGBAF32 = SkColorType::kRGBA_F32_SkColorType as _,
-    R8G8UNorm = SkColorType::kR8G8_unorm_SkColorType as _,
-    A16Float = SkColorType::kA16_float_SkColorType as _,
-    R16G16Float = SkColorType::kR16G16_float_SkColorType as _,
-    A16UNorm = SkColorType::kA16_unorm_SkColorType as _,
-    R16G16UNorm = SkColorType::kR16G16_unorm_SkColorType as _,
-    R16G16B16A16UNorm = SkColorType::kR16G16B16A16_unorm_SkColorType as _,
-    SRGBA8888 = SkColorType::kSRGBA_8888_SkColorType as _,
-    R8UNorm = SkColorType::kR8_unorm_SkColorType as _,
-}
-
-native_transmutable!(SkColorType, ColorType, color_type_layout);
-
-impl ColorType {
-    // error[E0658]: dereferencing raw pointers in constants is unstable (see issue #51911)
-    /*
-    pub const N32 : Self = unsafe {
-        *((&SkColorType::kN32_SkColorType) as *const _ as *const _)
-    };
-    */
-
-    pub fn n32() -> Self {
-        Self::from_native_c(SkColorType::kN32_SkColorType)
-    }
-
-    pub fn bytes_per_pixel(self) -> usize {
-        unsafe {
-            sb::SkColorTypeBytesPerPixel(self.into_native())
-                .try_into()
-                .unwrap()
-        }
-    }
-
-    pub fn is_always_opaque(self) -> bool {
-        unsafe { sb::SkColorTypeIsAlwaysOpaque(self.into_native()) }
-    }
-
-    pub fn validate_alpha_type(self, alpha_type: AlphaType) -> Option<AlphaType> {
-        let mut alpha_type_r = AlphaType::Unknown;
-        unsafe {
-            sb::SkColorTypeValidateAlphaType(self.into_native(), alpha_type, &mut alpha_type_r)
-        }
-        .if_true_some(alpha_type_r)
-    }
-}
 
 pub use skia_bindings::SkYUVColorSpace as YUVColorSpace;
 variant_name!(YUVColorSpace::JPEG, yuv_color_space_naming);
